delete unnecessary classes
